callback_plugins: init '_ansible_no_log' in plugin (#229)
The `_ansible_no_log` variable does not appear to be defined by
default by Ansible.  This was causing the callback plugin to suppress
any error messages about other undefined variables in a task, as
described in #226.

This change causes the variable to be initialized to `False` if the
variable is found to be undefined.  Doing so allows the other
undefined errors to bubble up.

Closes #226