Save original deployment info for later use (#225)
* Save original deployment info for later use

It's possible that through the course of testing that the original
deployment information is lost (more acccurately, not easily
available) via upgrades/rebases.  In order to return the host under
test to its original state, we need a way to retain the original
deployment information to be recalled later.

This problem is solved by writing out the deployment information
(`checksum`, `refspec`, and `remote`) to an INI-style file in the
local playbook directory.  This file can then be read in by another
role that will setup variables to be used by the `rpm_ostree_rebase`
role.

The solution is admittedly a bit specific to this problem/variables,
but I think could be modified in the future to be a bit more
versatile.

* Archive original deployment info to YAML file

This commit is a re-work of my the first approach to saving off the
original deployment info.  Here the information is templated out to a
YAML file and then loaded dynamically later on.

A generic role called `load_variables` has been provide that can be
used to load in variables from a YAML file.

This also removes the too-specific `original_refspec_set_fact` role.

* fixup! Archive original deployment info to YAML file