tests: drop 'rpm_signature_verify' after upgrade (#235)
At the beginning of the `improved-sanity-test`, we run the
`rpm_signature_verify` check to ensure that the supported streams are
checked to have the correct RPM signatures.  We skip this check on
certain streams which are known not to always have RPMs signed by a
single key.

Because the upgrade is performed against a local branch created from
the original commit, the content of the local branch is already well
known and has been checked for RPM signatures.  Thus we can skip the
RPM signature check for the second time.