roles: allow 'kubernetes_setup' to work with kube 1.5 and 1.6 (#248)
* roles: allow 'kubernetes_setup' to work with kube 1.5 and 1.6

This changes how the `kubernetes_setup` role pulls in containerized
kube images and which JSON files get dropped into the
`/etc/kubernetes/manifests` directory.  We inspect which version of
the `kubelet` is on the host and then make the necessary decisions
based on that.

Closes #246

* roles: use templates for pods

Rather than keeping two (or more) versions of the pod manifests, we
can use the `template:` function to write out the manifests
conditionally.