tests/rpm-ostree: don't try to 'rpm-ostree reload' on F26 (#264)
* tests/rpm-ostree: don't try to 'rpm-ostree reload' on F26

I had hoped that the latest two week release of F26 (26.141) would
have picked up the version of `rpm-ostree` (2017.9) which included a
fix for projectatomic/rpm-ostree#976, but that version was stuck in
Bodhi purgatory waiting for karma and missed the compose.  :rage:

Because that issue is impacting some of our PR testing, this change
temporarily switches to restarting the `rpm-ostreed` service for the
F26 case.  Once the *next* two-week compose is released, we will
should revert this change.

* fixup! tests/rpm-ostree: don't try to 'rpm-ostree reload' on F26