roles: refactor 'rpm_ostree_status*' to use json_query filter (#244)
* roles: refactor 'rpm_ostree_status*' to use json_query filter

This attempts to streamline the `rpm_ostree_status` role by employing
the `json_query` filter.  We can query the `deployments` from the
output of `rpm-ostree status --json` for the correct deployment which
is booted or not booted.  In the case of a single deployment, we
default to setting the `ros_not_booted` to false.

The `rpm_ostree_status_verify` has been converted to use the
json_query filter as well.  In this role, the `rpm_ostree_status` is
symlinked into the directory, so it can be included directly.  This
was done because we want to have that role run every time the 'verify'
role is called.  (I tried to use the meta dependencies, but it would
only run once per playbook if the 'verify' role was called multiple
times)

I also snuck it a little fix to the `admin-unlock` test that I noticed
while checking to see if the all the tests still worked.

Closes #230

* roles: enable allow_duplicates for rpm_ostree_status* roles

This changes the `rpm_ostree_status*` roles to be used more than once
in a playbook via the `allow_duplicates` toggle.

The side-effect is that the tests that use those roles multiple times
have to use them as roles or via `include_role:`.  So this change also
modifies the existing tests to adapt to this new requirement.

* tests: grab commits at beginnig of each 'rpm-ostree' test

We can run into trouble when running the `rpm-ostree` test if the
stream under test gets a new commit during the course of the test.  To
mitigate this, we pull the `HEAD` and `HEAD-1` commits before running
the bulk of each test.  This way the test will use those two commits
during the test, rather than getting an updated/new commit in the
middle of things.

* roles: enable reboot role to run multiple times

Flip the switch on 'allow_duplicates' so that the role can be used
multiple times in a single playbook.

* tests: use the 'reboot' role

Update the two tests to use the `reboot` role rather than explcitly
including the file.

Additionally, change the `rpm-ostree` test to use the parent of the
deployed commit as a stand-in for the `HEAD-1` commit.  The `HEAD-1`
commit that is fetched from the remote can be different from the
parent of the deployed commit, so this avoids any changes to the
remote that might upset the test.