tests: install a different package on CentOS (#261)
* roles: improve ability to check for un/installed packages

The `rpm_ostree_install*` and `rpm_ostree_uninstall*` roles had some
limitations when first created.  Namely, only being able to be used
once per playbook and being unable to check for a binary that did not
match the package name.

This changes the roles to use the `allow_duplicates` boolean, so that
they may be used multiple times in the same playbook.  The `*verify`
roles have also been changed to allow for checking for a different
binary name than what is used by the package.  For example, the `ntp`
package installs an `ntpd` binary, whereas the `httpd` package installs
an `httpd` binary.

* tests: handle package layering on CentOS

In #254, it is noted that the package layering of `httpd` isn't
working on CentOS because the `centos-logos` dependency is trying to
layer files on `/boot`.  Until CentOS gets a version of `rpm-ostree`
that supports layering on `/boot` (see projectatomic/rpm-ostree#969),
we need to use a different package to test package layering in CentOS.

This change brings in a conditional installation of `ntp` in the
CentOS case, until we have a newer version of `rpm-ostree` to use.

* test/i-s-t: use conditionl vars from vars files

This changes how the variables used by the `improved-sanity-test` are
loaded in each playbook section.  Commonly used variables are now
found in `vars/common.yml` and are loaded for each playbook section.
Additional variables that may change per OS platform, are broken out
into separate files.

By supplying a list of files to the `vars_files:` statement, Ansible
will try to import each file in the list and stopping when a file is
found.[0]  This allows us to override any variables per OS while
maintaing sensible defaults.

[0] http://docs.ansible.com/ansible/latest/playbooks_conditionals.html#conditional-imports

* roles/tests: handle shared variable namespace

When running a role multiple times in a playbook, we can run into
problems with variables since they all live in the same namespace.
Notably, when you run a role multiple times, variables that are
optional but have been defined, will carry over to the next execution
of the role.  Or if two roles have the same variable name, you can get
values stomping on each other.

This commit tries to alleviate some of these problems by updating the
variable names in the roles to be unique to the role.  Additionally,
when using roles multiple times in playbooks, both required and optional
variables are explicitly defined to avoid any overwriting.