roles: print matching lines from journal (#259)
* roles: print matching lines from journal

Previously, when the `journal_search` role would fail, it would only
print a message saying that the matching line had been found.  This is
only partially useful if you want to see the actual matching line in
the test output.

This changes the `journal_search` role to crudely print the matching
lines by way of running a `grep` against the output of `journalctl`
after a match has been found.  The reason for doing this operation
after the match has been found is because the command `journal | grep
<string>` may end up in the journal if it is run with elevated
privileges via Ansible (i.e. using `sudo`)

Closes #253

* journal_search: improve detection of messages in journal

We can exclude some of the messages in the journal that are generated by
running Ansible tasks/playbooks by just piping the output into an
inverse `grep` search.  This improves the reliability of the messages
getting matched by the role.

Additionally, this introduces support for passing extra arguments to
the `journalctl` command.  This allows users to be even more specific
with the journal messages they want to match; for example, matching
AVC denials using the `_TRANSPORT=audit` field.

* journal_fatal_msgs: be more restrictive when matching

We only want to match these messages that come from certain
transports, so utilize the `extra_args` variable to zero in on
specific messages that come from certain transports.