i-s-t: use diff RPM for 'rpm_install' on Fedora (#271)
* i-s-t: use diff RPM for 'rpm_install' on Fedora

Previously, we were getting away with using `epel-release` as a
generic RPM duing the `rpm_install` role in the i-s-t.  But the latest
version of the RPM now has a `Conflicts: fedora-release` and thus we
can't install it on Fedora-based AH systems anymore.

This change conditionalizes the RPM we use during the `rpm_install`
role.  On Fedora, I chose the current stable version of `ltrace` and
stuck with `epel-release` for everything else.

* fixup! i-s-t: use diff RPM for 'rpm_install' on Fedora