roles: adjust retries on 'redhat_subscription' and 'redhat_unsubscribe' (#275)
* roles: adjust retries on 'redhat_subscription' and 'redhat_unsubscribe'

Our internal jobs hit some network timeouts while doing an
'unsubscribe' with `subscription-manager`.  We already had some
retries for `redhat_subscription`, but were lacking them on the
`redhat_unsubscribe`.  This change implements the retries for
'unsubscribe' and bumps up the retries on 'subscription'

* redhat_subscription: add more retries

It looks like `subscription-manager` talks to the network for almost
every operation, so let's add some retries to each time we use it.

* fixup! redhat_subscription: add more retries