Use virtualenv for PAPR (#279)
Resolves issue #262.  The latest version of Ansible was being installed
by PAPR but A-H-T currently supports 2.2.x.  Use virtualenv to install
the supported version of Ansible for A-H-T.

While setting up virtualenv, the difference between CentOS and Fedora
containers was causing issues with the required rpms and the handling
of unset environment variables.  This PR makes all the tests run from
a Fedora 26 container to make the setup consistent across all the CI
streams.