rpm_install: use dummy RPM to support cross-platform installs (#282)
* rpm_install: use dummy RPM to support cross-platform installs

Previously we had been able to use the `epel-release` RPM as a way to
test RPM installs when a host had used `ostree admin unlock`.  But the
`epel-release` RPM was changed to conflict with the `fedora-release`
RPM, meaning we were not able to use a single RPM across all
platforms.

Instead of trying to setup conditions for when to use `epel-release`
and when not to, we are now able to use a 'dummy' RPM that will
install across all platforms.  (Thanks @jlebon!)

This change implements the use of the 'dummy' RPM in the tests that
use the `rpm_install` role.

Additionally, I snuck in some minor changes to the `epel_rpm_url`
role:  updating the base EPEL URL and improving the `sed` regex for
readability.

* rpm_install: implement installing RPM located on host

The `rpm_install` role currently operates by installing an RPM from a
URL provided.  With the addtion of the dummy RPM in #281, we can copy
the RPM to the host and perform the `rpm -i` using the file copied to
the host.

This change implements the ability to copy an RPM to the host and then
installing said RPM directly.  It retains the ability to install from
a URL, as that is still useful (and used) in other tests.

* fixup! rpm_install: implement installing RPM located on host

* fixup! rpm_install: implement installing RPM located on host

* fixup! rpm_install: implement installing RPM located on host

* fixup! rpm_install: implement installing RPM located on host