Add RPM version retrieval (#291)
* Add RPM version retrieval

Sometimes functionality may differ in an RPM released in one atomic
host versus another.  Conditionals must be setup to handle the
differences.  Until now, this was done in the individual test/role.
This commit centralizes that functionality into one role so the
test can reference the same dictionary to get RPM version information.

rpm_version is a base role that will get the RPM versions from a list
of RPMs.  atomic_host_versions gathers version information for a default
set of RPMS (or user specified set of RPMs) and can also optionally fail
if minimum versions of RPMs are not met.

Removed runc_version_check role because it is not used and obsoleted
by the addition of these roles.

* fixup! Add RPM version retrieval