use 'jq' to filter out noisy keys from JSON output (#301)
* roles:  new role to install 'jq' utility

Having the `jq` utility on the host allows us to manipulate JSON
output more easily than trying to do it in Ansible.  This introduces a
role to check for the utility and install it if it does not exist.

* roles/rpm_ostree_status: exclude noisy keys

In the newer versions of `rpm-ostree`, the RPM DB package list is
included in the commit metadata.  This causes the output of
`rpm-ostree status --json` to be very noisy when run via Ansible (i.e.
100s of package key/values are now in the output).

This attempts to work around the noise by running the output through
`jq` and deleting the offending keys.

* Revert "roles:  new role to install 'jq' utility"

This reverts commit a6fe955aa6230b3b2e3dc01cf22d2e3ba6947f79.

* tools:  add Dockerfile for tools container

This Dockerfile will build a small tools container based on the
`fedora-minimal` image.  This first version only installs the `jq`
utility.

I've initially built the image in my own namespace, but we can make it
an automated build from this repo after this change gets merged.

* fixup! roles/rpm_ostree_status: exclude noisy keys

* roles/docker_pull: add retries

Network conditions can occasionally cause pull operations to fail, so
lets retry a few times to see if we can work around them.