Comply to YAML spec where possible (#302)
* Comply to YAML spec where possible

This commit fixes minor issues in the atomic host tests to comply
to the YAML spec.  The issues fixed were found using yamllint.

There were some instances of line length that could not be fixed
(i.e. Ansible when statements like  "'foo' in bar"). Using > or |
would break the logic so the line length had to be preserved.

* fixup! Comply to YAML spec where possible

* fixup! Comply to YAML spec where possible

* fixup! Comply to YAML spec where possible