bdsf: utilize the tools container to hide noisy JSON keys (#327)
Similar to the `rpm_ostree_status` role, when we output the status in
JSON format, one of the keys has a list of all the packages that are
included in the ostree commit.  In f6a7086, we introduced the use of a
tools container to pipe the JSON output through `jq` to hide the noisy
keys.  This is basically the same implementation.

(We probably should consolidate the two roles, but that is for another
PR...)