k8s-cluster: fix test for F27AH (#335)
* Install k8 worker nodes and etcd system containers for Fedora 27AH

This is required, as from Fedora27AH, k8, flannel and etcd are no
longer part of the Host's image.

Signed-off-by: Kedar Bidarkar <kbidarka@redhat.com>

* k8s-cluster: fix test for F27AH

In Fedora 27 Atomic Host, all of the `kubernetes` related packages
were removed from the host.  The replacement functionality is now
provided by system containers.

This set of changes reworks the `kubernetes_setup` role to utilize the
new system containers to bring up a single node cluster and then
proceed with the remaininder of the `k8s-cluster` test.

NOTE:  the system containers are only used in the case where the
`kubernetes` packages are not present on the host.  For other
platforms where they are present on the host, the original method of
using manifests continues to be used.

Many thanks to Kedar Bidarkar for starting this work!

* kubernetes_set_facts: new role to determine kube facts

This adds a utility role that is meant to be used as a dependency on
other roles that need to determine certain pieces of kubernetes
information, like version, names of containers, or whether to use
system containers.

* fixup! k8s-cluster: fix test for F27AH

* kubernetes_set_facts: move more logic into role

This commit moves the logic for determining `kubectl_path` into the
`kubernetes_set_facts` role.  The roles which utilize `kubectl_path`
have also been updated to depend on the `kubernetes_set_facts` role.

I also did some slight rework of the `k8_remove_all` role to handle
more of the services that are installed when using system containers.