roles: convert docker roles to be more generic (#331)
* roles: convert docker roles to be more generic

We have a number of specialized roles that are used in the build, run,
and removal of an `httpd` container.  While this has worked thusfar,
it limits the flexibility around the build/run/remove operations of
Docker containers.

This change modifies the existing roles to be more generic and allows
us to use them with more flexibility.

As part of the changes, a new role named `check_open_port` was added
to replace some functionality that was removed from an exiting role.

Finally, the existing tests that used the roles were updated to use
the new role names.

* fixup! roles: convert docker roles to be more generic

* fixup! roles: convert docker roles to be more generic

* fixup! roles: convert docker roles to be more generic

* docker_rmi: rework to support tagged images

Previously, if a tagged image was passed to the role, it could fail
because the role would search for the string in the output of `docker
images`.  But the output of `docker images` uses a table format and
separates the image name from the tag.

This change reworks the role to specifically select the image name
from the output of `docker images` and then decides if output returned
is what is expected.  If we are looking for the image to be present,
`docker images` will return 2 lines of stdout.  Whereas, if we are
looking for the image to be missing, `docker images` will only return
1 line of stdout.

Additionally, the `docker rmi` command will now always use the `-f`
option to forcefully remove the image.