k8-cluster: use kubeconfig (#357)
The k8-cluster test was failing in the system container use case.  The
web and db containers weren't being created because no nodes were
available.  I was able to root cause the issue to the deprecation of
the --api-servers option.  Kubernetes 1.9 uses a kubeconfig file for
the node to determine the api server.  I used a kubernetes PR [0] as
reference for the creation of the kubeconfig file and then tacked on
the --kubeconfig option to the KUBELET_ARGS.

[0] https://github.com/kubernetes/contrib/pull/2863/files