README: new dashboard for 'next' tier tests (#310)
In the internal Red Hat infrastructure, we are now running a select
number of 'next' tier tests from the repo after a successful run of
the `improved-sanity-test`.  This commit creates a separate dashboard
that tracks the status of each of these tests.

The 'next' tier results are presented in a single 4 column table with
delineations made using empty rows.  This was done to allow for all
the badges to be legible when reviewing the results.