system-containers: use FQ container name (#360)
When a fully qualified container name is not used, skopeo will use
the default container registry and use a short shorname.  In this
case gscrivano/hello-world shows up as hello world in the atomic
images list output.  The system-container looks for the short name
that was passed in when installing the system container so the test
ends up failing.  The simple fix is to use fully qualified container
names.