i-s-t: add test for non-priv systemd container (#333)
* i-s-t: add test for non-priv systemd container

Since it is possible to run systemd in a container without the
`--privileged` flag, we decided we should have a test for this going
forward.

This change adds in the files necessary to build an `httpd` container
that uses `systemd` as its init mechanism.  Additionally, we add the
building/running of said container to the `improved-sanity-test`.

* i-s-t: configure SELinux boolean for systemd container test

In order to run systemd in a container, we need to enable an SELinux
boolean on the host.  This change does the needful for said boolean
(container_manage_cgroup).

* i-s-t: use additional tags around systemd container test

In order to have more control over which tests are run/excluded, this
adds additional tags around the roles for the non-privileged systemd
httpd container test.