docker-swarm: disable live-restore
The live-restore option was enabled for fedora 28 which caused a
failure with the docker-swarm test.  The default behavior for docker
is to shutdown all running containers when the docker daemon
terminates.  The live-restore option allows containers to continue to
run when the docker daemon is unavailable; however, docker swarm
cannot run with live-restore enabled because it needs to manage the
lifecycle of the containers.

This commit disables live-restore that is enabled through
/etc/syconfig/docker.

Closes issue #370