tree: initial changes to support Fedora STI/STR (#368)
* i-s-t:  initial changes to support Fedora STI/STR

See the following for some details:

https://fedoraproject.org/wiki/CI/Tests
https://fedoraproject.org/wiki/CI/Standard_Test_Roles

These changes allow the `improved-sanity-test` playbook to be invoked
using the same pattern as described in the docs above.

The notable change is the value to `hosts:` which will try to find a
target under the `subjects` group (which is how the inventory file is
built via the STI).  But defaulting to `all` retains backward
compatibility with the original way of running the playbook.

* reboot: modify to support Fedora STR

Additional details:

https://fedoraproject.org/wiki/CI/Tests
https://fedoraproject.org/wiki/CI/Standard_Test_Roles

When running any of the 'a-h-t' tests against an 'atomic' test subject
as defined by the Standard Test Interface, the harness will boot a VM
using `qemu` and fowards port 22 to a non-standard port (typically
I've seen 2222).  The side-effect of this is that the non-standard
port is always up and open, unless the VM has been torn down.  This
does *not* include when the host is rebooted.

This behavior causes our original methodology of waiting for the port
to go down and then come back up to be invalid.

I borrowed the idea of checking the `bootid` from the `rpm-ostree`
tests.  I also had to add in the ability to configure the `delay` on
the `wait_for` because occasionally the VM booted by the STI wouldn't
be ready within the default 30s.  Thankfully, this approach is
backwards compatible with the previous ways we have run the playbooks.

I'm still not 100% certain this is going to work all the time, but it
is a step in that direction.

* reboot: use search_regex to improve resiliency

@mike-nguyen found this post that talks about a similar problem we are
having with hosts started via the Fedora STI, specifically when the
standard SSH port is forwarded to a non-standard port:

https://blog.jeffli.me/blog/2017/05/19/a-pitfall-of-vagrant-when-using-ansibles-wait-for-module/

Using the `search_regex` function of `wait_for` proved to be useful in
this scenario, as we can leave the default delay of 30s intact and
don't have to bump that up.  (Which is effectively just sleeping the
whole execution for that amount of time.)

* fixup! reboot: use search_regex to improve resiliency