tests: use 'include_tasks' for Ansible 2.4 support
The use of `include` to pull in task files is deprecated in newer
versions of Ansible and will eventually be removed.  This updates the
tests to use `include_tasks` where `include` was removed.

Additionally, I changed a portion of the `pkg-layering` test where it
was looking for a string in the sdterr to use a logical `AND` instead of
a big ugly string.  Ansible was issuing a warning about the use of the
"{{ var }}" syntax in a `when` statement; changing to a logical `AND`
silenced the warning and achieves (nearly) the same result.