Allow all tests to run even when one fails
Ansible fails fast by default.  With the introduction of block and
rescue, ansible can catch failures and continue on its merry way.  We
can leverage this functionality by using using a simple role wraps an
include_task in a block/rescue and capturing results in a dictionary.
This strategy requires the tests to be broken out into individual task
yml files.  This could also help alignment with the Standard Test
Interface work which uses an entry point script to run tests.