rpm-ostree: add override tests (#374)
* rpm-ostree: add override tests

The rpm-ostree override functionality has been out of experimental for
a while now so this commit adds tests for replace, remove, and reset.

These tests replace and remove packages that are common to RHEL,
Fedora, and CentOS Atomic Hosts.  The goal was to test against an
ostree commit release without modification.  The caveat to this approach
is that the package set may change and we will have to find another
common package to use.  If this becomes too troublesome, we can adopt
the strategy that the rpm-ostree unit tests uses and create a new
branch with a known rpm to replace and remove.

Closes #345

* fixup! rpm-ostree: add override tests