Resolve Ansible 2.4  import_role variable issue (#389)
There are two issues encountered after the Ansible 2.4 change.  The
first issue is the rpm_ostree_rebase role.  The refspec variable
inside the role seems to get populated with refspec somewhere up
before the rebase call and causes a failure.  This was fixed by using
a prefixed variable name (ror_refpsec) which is a best practice
anyhow.

The second issue is a little more interesting.  After calling the same
role (rpm_ostree_install_verify) multiple times in the rpm-ostree test,
the import_role doesn't seem to be respecting the variable that is
being passed in.  Instead it is using a value from somewhere up before
the same role is called.  This leads to the conclusion that prefix in
the rebase variable name above may not work in all cases.  This was
fixed by using an include_role which is a dynamic include.  We were
trying to avoid using include_role too often because there is a
recursion depth limit if used too many times.