roles: don't fail on SELinux failures; record them to files (#384)
A common problem when runing the 'sanity' test is that it can fail
very early because a stream has an SELinux denial immediately upon
booting the host.  While it is helpful to detect this kind of denial,
failing the whole test suite near the beginning could let other
problems slip by.

This change attempts to avoid the 'early failure' scenario by
capturing the various SELinux failures and dumping them to files on
the host where the playbook is running.  I'm open to changing how the
files are named or what not.

I've made a few assumptions about how this works because we are
basically the only ones using these roles/tests.  To everyone else,
YMMV.