admin-unlock: continue to run on failure (#391)
* admin-unlock: continue to run on failure

Instead of failing early (the default behavior for Ansible), this PR
uses Ansible's block/rescue to catch errors and continue execution.

There are a few major changes in the structure of the test.
- Each tests is broken into individual yaml files as set of tasks
- Each test that will continue on failure is added as an include_task
in the main.yml file in a block/rescue.
- Each test that we want to fail fatally will not be in a block/rescue
- Each test will capture the result and result details in a list of
dictionaries that can be evaluated at the end of the main.yml test
- When an exception is caught in a block/rescue the failed task
details is stored in ansible_failed_task and ansible_failed_result

The goal was to organize the tests better simplify adding new tests to
the test suite.  main.yml is the test suite and individual testa are
in their own yaml file.  main.yml is much uglier and repetitive due to
an Ansible bug [0] where ansible_failed_result and ansible_failed_task
is undefined if used within a role or with includes.  Ideally, there
would be a role that will include_tasks taking a yaml file is a
parameter and do all the test results tracking.  Once the bug [0] is
fixed, a test can be included by just calling the role.

[0] https://github.com/ansible/ansible/issues/29047

* fixup! admin-unlock: continue to run on failure