Adapt the 'rpm-ostree' test suite to continue on failure (#404)
* gitignore: ignore .log files

* rpm-ostree: use 'continue on fail' model

This changes the `rpm-ostree` test suite to use the 'continue on
failure' model.

Partially addresses #375

* rpm-ostree/livefs: use global variable name

* rpm-ostree/compose: make process more resilient

This worked well on a pristine system, but I was getting tripped up
when re-using a test sytsem.  This cleans up some of the transient
artifacts that can be generated by the compose.

Additionally, I added a `wait_for` to make sure the HTTP server
starts.

* rpm-ostree/livefs: use include_role in a few places

We were seeing instances where variable values passed to `import_role`
where not being respected each time it was called.  I thought that
each time an 'import' was done the changed variables would be
respected, but that doesn't appear to be the case.

Using `include_role` will allow us to dynamically load in variables
each time the role is used.  This is kind of troubling as we made a
large switch to using `include_role` in most places, but we've not run
into any problems like this yet.

* rpm-ostree: explicitly check for failure

This is a forward-port of functionality from #405 which corrects an
assumption about how Ansible was handling failures.

* move 'ostree admin status' to sanity test

In #381, a sanity check for `ostree admin status` was added to the
`rpm-ostree` test suite.  It is probably better suited to the 'sanity'
test suite so we are able to catch any regressions earlier in the
testing cycle.

I created a dumb role for the `ostree admin status` command since that
was easier than trying to wedge it into an existing role.