rpm-ostree: workaround ansible variable issue
We have been sporadically hitting an issue with Ansible where the
arguments passed in are not overwriting the previous values for those
variables.  This may be an Ansible bug [0].

The only way we have been able to mitigate this is by using
include_role instead of import role.  So why not use include_role
everywhere?  Ansible has a recursion depth limit that is hit when
include_role is used often.

Another change that was needed was to make the binary check optional
since some RPMs don't exactly have a binary that can be checked with
command -v.

[0] https://github.com/ansible/ansible/issues/37787